Fix WeatherInfo saving data in same datastore with different format
Two datastores were being saved in the same file, and this miraculously worked until now, but it's not the intended way to handle this.
I don't think there is a proper way to do a migration from a reused file, so I am just changing the file name.
This means that the weather city the users set will be lost, but hopefully it's not a big issue.